pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-venv"
authors = [
    {name = "Ash Berlin-Taylor", email = "ash_github@firemirror.com"},
    {name = "Ian Buss", email = "ian.buss@astronomer.io"},
    {name = "Kaxil Niak", email = "kaxil@astronomer.io"},
]
license = {text = "Apache License 2.0"}
description = "Easily create and use Python Virtualenvs in Apache Airflow"
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]
urls = {Homepage = "http://astronomer.io/"}
dynamic = ["version"]
requires-python = "~=3.7"
dependencies = ["apache-airflow>=2.4"]

[project.optional-dependencies]
tests = [
    "pytest>=7.0",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "astronomer.providers.venv:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = {attr = "astronomer.providers.venv.__version__"}

[tool.setuptools.packages.find]
where = ["provider"]

[tool.ruff]
fix = true
line-length = 110
target-version = "py37"
extend-select = [
    "PGH", #pygrep-hooks
    "UP", # Pyupgrade
    "RUF100", # Unused noqa (auto-fixable)
    "I",
]

[tool.pytest.ini_options]
addopts = "--tb=short -v"
filterwarnings = [
    "ignore::DeprecationWarning:apispec.utils",
    "ignore::DeprecationWarning:connexion.decorators.validation",
    "ignore::DeprecationWarning:marshmallow_sqlalchemy.convert",
    "ignore::DeprecationWarning:flask_sqlalchemy",
    "ignore::DeprecationWarning:flask_appbuilder.filemanager",
    "ignore::DeprecationWarning:flask_appbuilder.widgets",
    "ignore::DeprecationWarning:pkg_resources",
]
